add server package
